* Исправление ошибок при передаче блоков данных на ВУ
- Исправлен подсчет числа свободных и занятых блоков
  в контейнере блоков данных
- Отладочный вывод при передаче данных на ВУ

+++ src/comm/putx.c
@@ -79,7 +79,7 @@ void putdwx_zs(const unsigned long w)
 }                
 
 //-------------------------------------------------------------------
-void putd(int d)
+void putd(short d)
 {
 	unsigned char buf[7]; 
 	register unsigned char* p = &buf[sizeof(buf)-1];
